Update dependencies in package.json
