+++ api/routes/user.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const router = express.Router();
+
+// Fetch user profile
+router.get("/profile", (req, res) => {
+    // get user profile logic
+});
+
+module.exports = router;
