--- routes/course.js
@@ -1,14 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-// Get course details by ID
-router.get("/:courseId", (req, res) => {
-    // handle getting course by ID
-});
-
-// Add new course
-router.post("/add", (req, res) => {
-    // handle adding new course
-});
-
-module.exports = router;
