--- routes/user.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-// Fetch user profile
-router.get("/profile", (req, res) => {
-    // get user profile logic
-});
-
-module.exports = router;
